fix(app): address rules search review
