+++ tests/app-tests/config.yaml
@@ -31,6 +31,7 @@ executionOrder:
         - 25.rules-crud.flow
         - 26.rules-application.flow
         - 27.mcc-default-category.flow
+        - 28.rules-search.flow
 
 platform:
     ios:
